Legger til mappe for intro-oppgavene og oppdaterer readme
Renamer endel mapper/filer
