+++ src/main/kotlin/no/bekk/game/Config.kt
@@ -1,4 +1,4 @@
-package org.veiset.libgdx
+package no.bekk.game
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration
