+++ src/main/kotlin/no/bekk/game/Main.kt
@@ -1,7 +1,5 @@
-package org.veiset.libgdx.solution
+package no.bekk.game
 
-import com.badlogic.gdx.Gdx
-import com.badlogic.gdx.Input
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication
 import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer
@@ -9,105 +7,21 @@ import com.badlogic.gdx.math.Vector2
 import com.badlogic.gdx.utils.TimeUtils
 import ktx.graphics.rect
 import ktx.graphics.use
-import org.veiset.libgdx.*
-import org.veiset.libgdx.utils.Rectangle
-import org.veiset.libgdx.utils.x
+import no.bekk.game.utils.Rectangle
+import no.bekk.game.utils.x
 
 fun main() {
-    LwjglApplication(AppRunner { DodgeFallingSqueresSolution() }, config)
+    LwjglApplication(AppRunner { DodgeFallingSqueres() }, config)
 }
 
-class DodgeFallingSqueresSolution: AppModule {
+class DodgeFallingSqueres: AppModule {
     private val shapeRenderer = globals.shapeRenderer
     private var lastBlockSpawnTime = TimeUtils.millis()
-    private var movementSpeed = 500f
-    private var blocksFallSpeed = 250f
     private var player = Rectangle(
-        position = EngineConfig.width / 2 x 200,
-        size = 20 x 20
+        position = Vector2(EngineConfig.width / 2f, 200f),
+        size = Vector2(20f, 20f)
     )
     private var blocksToDodge: List<Rectangle> = emptyList()
-
-    /**
-     * In this method we will implement player movement.
-     * You can check for keypresses using `Gdx.input.isKeyPressed`
-     * for instance you can check for UP-arrow: `Gdx.input.isKeyPressed(Input.Keys.UP)`.
-     *
-     * @param delta The time since last update. This is used to make the movement-speed
-     * not dependant of the update-rate.
-     */
-    private fun handlePlayerMovement(delta: Float) {
-        if (Gdx.input.isKeyPressed(Input.Keys.D) || Gdx.input.isKeyPressed(Input.Keys.RIGHT)) {
-            player.position.x += delta * movementSpeed
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.A) || Gdx.input.isKeyPressed(Input.Keys.LEFT)) {
-            player.position.x -= delta * movementSpeed
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.W) || Gdx.input.isKeyPressed(Input.Keys.UP)) {
-            player.position.y += delta * movementSpeed
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.S) || Gdx.input.isKeyPressed(Input.Keys.DOWN)) {
-            player.position.y -= delta * movementSpeed
-        }
-
-        if (player.position.x < 0) { player.position.x = 0f }
-        if (player.position.x > EngineConfig.width) { player.position.x = EngineConfig.width.toFloat() }
-
-        if (player.position.y < 0) { player.position.y = 0f }
-        if (player.position.y > EngineConfig.height) { player.position.y = EngineConfig.height.toFloat() }
-    }
-
-    /**
-     * Checks if the game should spawn new blocks. You can use the given timeSinceLastBlock
-     * to determine wheter to spawn a new block or not.
-     *
-     * @param timeSinceLastBlock The time since the last block was spawned in milliseconds.
-     * @return true if the game should spawn new blocks, false if not
-     */
-    private fun shouldSpawnNewBlocks(timeSinceLastBlock: Long): Boolean {
-        return timeSinceLastBlock > 100
-    }
-
-    /**
-     * Here we will implement logic for spawning new blocks that the player
-     * has to dodge. The blocks should start at top of the screen at a random
-     * X-coordinate. The the blocks should be added to the list `blocksToDodge`.
-     */
-    private fun spawnNewBlocks() {
-        val position = Vector2(
-            (Math.random() * EngineConfig.width).toFloat(),
-            EngineConfig.height + 20f
-        )
-        val block = Rectangle(
-            position = position,
-            size = 20 x 20
-        )
-        blocksToDodge = blocksToDodge + listOf(block)
-    }
-
-    /**
-     * Moves all the blocks in `blocksToDodge` list.
-     *
-     * @param delta The time since last update. This is used to make the movement-speed
-     * not dependant of the update-rate.
-     */
-    private fun handleMoveBlocks(delta: Float) {
-        blocksToDodge = blocksToDodge
-            .map { it.move(0f x -blocksFallSpeed * delta) }
-    }
-
-    /**
-     * Removes all the blocks that are out of bounds. Since the blocks start at the top
-     * and fall downwards we can remove any blocks that are below the screen.
-     */
-    private fun removeBlocksOutOfBounds() {
-        blocksToDodge = blocksToDodge
-            .filter { it.position.y > 0 }
-    }
-
-    private fun playerIsColliding(): Boolean {
-        return blocksToDodge.any { it.isCollidingWith(player) }
-    }
 
     /**
      * This function should draw the player in the current position on the screen.
@@ -118,11 +32,75 @@ class DodgeFallingSqueresSolution: AppModule {
     }
 
     /**
+     * In this method we will implement player movement.
+     * You can check for keypresses using `Gdx.input.isKeyPressed`
+     * for instance you can check for UP-arrow: `Gdx.input.isKeyPressed(Input.Keys.UP)`.
+     *
+     * @param delta The time since last update. This is used to make the movement-speed
+     * not dependant of the update-rate.
+     */
+    private fun handlePlayerMovement(delta: Float) {
+
+    }
+
+    /**
+     * Checks if the game should spawn new blocks. You can use the given timeSinceLastBlock
+     * to determine wheter to spawn a new block or not.
+     *
+     * @param timeSinceLastBlock The time since the last block was spawned in milliseconds.
+     * @return true if the game should spawn new blocks, false if not
+     */
+    private fun shouldSpawnNewBlocks(timeSinceLastBlock: Long): Boolean {
+        return false
+    }
+
+    /**
+     * Here we will implement logic for spawning new blocks that the player
+     * has to dodge. The blocks should start at top of the screen at a random
+     * X-coordinate. The blocks should be added to the list `blocksToDodge`.
+     *
+     * Keep in mind that most things (like lists) are immutable in Kotlin. So we
+     * can't actually add anything to the list. Instead we want to make a new list.
+     */
+    private fun spawnNewBlock() {
+
+    }
+
+    /**
      * This function should draw all the blocks the player has to dodge on the screen.
      * Use the utility method `drawRectangle` to draw on the screen.
      */
     private fun drawAllBlocksToDodge() {
-        blocksToDodge.forEach { drawRectangle(it, Color.GRAY) }
+
+    }
+
+    /**
+     * Moves all the blocks in `blocksToDodge` list.
+     *
+     * @param delta The time since last update. This is used to make the movement-speed
+     * not dependant of the update-rate.
+     */
+    private fun handleMoveBlocks(delta: Float) {
+
+    }
+
+    /**
+     * Checks if the player is colliding with any of the blocks in the `blocksToDodge`-list.
+     * `Rectangle` has a utility method `isCollidingWith` that checks if 2 Rectangles are
+     * overlapping.
+     *
+     * @return true if the player is colliding with any blocks, false if not.
+     */
+    private fun playerIsColliding(): Boolean {
+        return false
+    }
+
+    /**
+     * Removes all the blocks that are out of bounds. Since the blocks start at the top
+     * and fall downwards we can remove any blocks that are below the screen.
+     */
+    private fun removeBlocksOutOfBounds() {
+
     }
 
     /**
@@ -132,8 +110,8 @@ class DodgeFallingSqueresSolution: AppModule {
      */
     private fun onGameLost() {
         player = Rectangle(
-            position = EngineConfig.width / 2 x 200,
-            size = 20 x 20
+            position = Vector2(EngineConfig.width / 2f, 200f),
+            size = Vector2(20f, 20f)
         )
         blocksToDodge = emptyList()
     }
@@ -149,7 +127,7 @@ class DodgeFallingSqueresSolution: AppModule {
         handlePlayerMovement(delta)
         if (shouldSpawnNewBlocks(TimeUtils.timeSinceMillis(lastBlockSpawnTime))) {
             lastBlockSpawnTime = TimeUtils.millis()
-            spawnNewBlocks()
+            spawnNewBlock()
         }
 
         handleMoveBlocks(delta)
@@ -165,6 +143,12 @@ class DodgeFallingSqueresSolution: AppModule {
         drawAllBlocksToDodge()
     }
 
+    /**
+     * Draws a filled rectangle on the screen.
+     *
+     * @param rectangle The rectangle to draw
+     * @param color The color of the rectangle
+     */
     private fun drawRectangle(rectangle: Rectangle, color: Color) {
         shapeRenderer.use(ShapeRenderer.ShapeType.Filled) {
             it.color = color
