+++ src/main/kotlin/no/bekk/game/examples/PerlinModule.kt
@@ -1,4 +1,4 @@
-package org.veiset.libgdx.examples
+package no.bekk.game.examples
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication
@@ -6,11 +6,11 @@ import com.badlogic.gdx.graphics.GL20
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer
 import ktx.graphics.color
 import ktx.graphics.use
-import org.veiset.libgdx.AppModule
-import org.veiset.libgdx.AppRunner
-import org.veiset.libgdx.config
-import org.veiset.libgdx.globals
-import org.veiset.libgdx.utils.math.Perlin
+import no.bekk.game.AppModule
+import no.bekk.game.AppRunner
+import no.bekk.game.config
+import no.bekk.game.globals
+import no.bekk.game.utils.math.Perlin
 import kotlin.random.Random
 
 fun main() {
