+++ src/main/kotlin/no/bekk/game/utils/Font.kt
@@ -1,4 +1,4 @@
-package org.veiset.libgdx.utils
+package no.bekk.game.utils
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Texture
