+++ src/main/kotlin/no/bekk/introduction/Intro.kt
@@ -1,0 +1,10 @@
+package no.bekk.introduction
+
+class Konsulent(val name: String)
+
+// Klasse konsulent
+
+fun main() {
+    // Print konsulent
+    // Flere statements kommentert ut
+}
